fix: 自身以外のブックを削除できるAPIの問題の修正

+++ server/services/book/destroy.ts
@@ -1,5 +1,7 @@
 import { FastifySchema } from "fastify";
 import { BookParams, bookParamsSchema } from "$server/validators/bookParams";
+import { Session, isUserOrAdmin } from "$server/utils/session";
+import bookExists from "$server/utils/book/bookExists";
 import destroyBook from "$server/utils/book/destroyBook";
 
 export const destroySchema: FastifySchema = {
@@ -7,13 +9,27 @@ export const destroySchema: FastifySchema = {
   params: bookParamsSchema,
   response: {
     204: { type: "null", description: "成功" },
+    400: {},
+    403: {},
+    404: {},
   },
 };
 
-export async function destroy({ params }: { params: BookParams }) {
+export async function destroy({
+  session,
+  params,
+}: {
+  session: Session;
+  params: BookParams;
+}) {
+  if (!session.user) return { status: 400 };
+
+  const found = await bookExists(params.book_id);
+
+  if (!found) return { status: 404 };
+  if (!isUserOrAdmin(session, { id: found.authorId })) return { status: 403 };
+
   await destroyBook(params.book_id);
 
-  return {
-    status: 204,
-  };
+  return { status: 204 };
 }
